requirements.txt
altgraph==0.17
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
idna==2.10
numpy==1.20.3
pefile==2021.5.24
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
pywin32-ctypes==0.2.0
requests==2.25.1
urllib3==1.26.5
